Revert "Add collections filters"
