+++ schema.gql
@@ -430,15 +430,6 @@ input CollectionsFilter {
   type: NftTypeEnum
 }
 
-input CollectionsSort {
-  field: CollectionsSortEnum = VERIFIED
-}
-
-enum CollectionsSortEnum {
-  TIMESTAMP
-  VERIFIED
-}
-
 input ConnectionArgs {
   """Paginate after opaque cursor"""
   after: String
@@ -775,13 +766,7 @@ type Query {
   auctionsSortByBids(filters: FiltersExpression, grouping: Grouping, pagination: ConnectionArgs): AuctionResponse!
   campaigns(filters: CampaignsFilter, pagination: ConnectionArgs): CampaignsResponse!
   collectionStats(filters: CollectionStatsFilter!): CollectionStats!
-  collections(
-    filters: CollectionsFilter
-    pagination: ConnectionArgs
-
-    """This sort is descending and will apply only to full list"""
-    sort: CollectionsSort
-  ): CollectionResponse!
+  collections(filters: CollectionsFilter, pagination: ConnectionArgs): CollectionResponse!
   featuredCollections(pagination: ConnectionArgs): CollectionResponse!
   featuredNfts(pagination: ConnectionArgs): AssetsResponse!
   marketplaceCutPercentage: String!
