+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config/dist';
 import { GraphQLModule } from '@nestjs/graphql';
 import 'reflect-metadata';
-import { CollectionsModuleGraph } from './modules/nftCollections/collections.module';
+import { CollectionModuleGraph } from './modules/nftCollections/collection.module';
 import { AssetsModuleGraph } from './modules/assets/assets.module';
 import { AuctionsModuleGraph } from './modules/auctions/auctions.module';
 import { OrdersModuleGraph } from './modules/orders/orders.module';
@@ -49,7 +49,7 @@ import { NotificationsModuleGraph } from './modules/notifications/notifications.
       },
     }),
     CommonModule,
-    CollectionsModuleGraph,
+    CollectionModuleGraph,
     AssetsModuleGraph,
     AssetHistoryModuleGraph,
     AuctionsModuleGraph,
