+++ src/crons/cache.warmer/cache.warmer.module.ts
@@ -1,9 +1,9 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { CommonModule } from 'src/common.module';
 import { AuctionsModuleGraph } from 'src/modules/auctions/auctions.module';
-import { CollectionsNftsCountRedisHandler } from 'src/modules/nftCollections/collections-nfts-count.redis-handler';
-import { CollectionsNftsRedisHandler } from 'src/modules/nftCollections/collections-nfts.redis-handler';
-import { CollectionsService } from 'src/modules/nftCollections/collections.service';
+import { CollectionsNftsCountRedisHandler } from 'src/modules/nftCollections/collection-nfts-count.redis-handler';
+import { CollectionsNftsRedisHandler } from 'src/modules/nftCollections/collection-nfts.redis-handler';
+import { CollectionsService } from 'src/modules/nftCollections/collection.service';
 import { PubSubListenerModule } from 'src/pubsub/pub.sub.listener.module';
 import { AuctionsWarmerService } from './auctions.warmer.service';
 import { CollectionsWarmerService } from './collections.warmer.service';
