+++ src/modules/common/filters/filtersTypes.ts
@@ -151,6 +151,43 @@ export class AssetsFilter {
 }
 
 @InputType()
+export class CollectionsFilter {
+  @IsOptional()
+  @Matches(RegExp(ADDRESS_RGX), { message: ADDRESS_ERROR })
+  @Field(() => String, {
+    nullable: true,
+    description: 'The owner of the collection',
+  })
+  ownerAddress: string;
+
+  @IsOptional()
+  @Matches(RegExp(ADDRESS_RGX), { message: ADDRESS_ERROR })
+  @Field(() => String, {
+    nullable: true,
+    description: 'The user that has create role',
+  })
+  creatorAddress: string;
+
+  @Field(() => Boolean, {
+    nullable: true,
+    description: 'Flag for can create or not on collection',
+  })
+  canCreate: boolean;
+
+  @IsOptional()
+  @Matches(RegExp(COLLECTION_IDENTIFIER_RGX), {
+    message: COLLECTION_IDENTIFIER_ERROR,
+  })
+  @Field(() => String, {
+    nullable: true,
+    description: 'Collection identifier',
+  })
+  collection: string;
+  @Field(() => NftTypeEnum, { nullable: true })
+  type: NftTypeEnum;
+}
+
+@InputType()
 export class CampaignsFilter {
   @Field(() => String)
   campaignId: string;
