+++ src/modules/nft-rarity/nft-rarity.module.ts
@@ -4,15 +4,15 @@ import { ElrondCommunicationModule } from 'src/common';
 import { CommonModule } from 'src/common.module';
 import { RarityUpdaterService } from 'src/crons/elastic.updater/rarity.updater.service';
 import { NftRarityRepository } from 'src/db/nft-rarity/nft-rarity.repository';
+import { CollectionModuleGraph } from 'src/modules/nftCollections/collection.module';
 import { AssetRarityInfoRedisHandler } from '../assets/loaders/assets-rarity-info.redis-handler';
-import { CollectionsModuleGraph } from '../nftCollections/collections.module';
 import { NftRarityComputeService } from './nft-rarity.compute.service';
 import { NftRarityService } from './nft-rarity.service';
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([NftRarityRepository]),
-    CollectionsModuleGraph,
+    CollectionModuleGraph,
     ElrondCommunicationModule,
     CommonModule,
   ],
