+++ src/modules/nftCollections/collection.module.ts
@@ -1,6 +1,6 @@
 import { forwardRef, Module } from '@nestjs/common';
-import { CollectionsQueriesResolver } from './collections-queries.resolver';
-import { CollectionsService } from './collections.service';
+import { CollectionsQueriesResolver } from './collection-queries.resolver';
+import { CollectionsService } from './collection.service';
 import { AssetsModuleGraph } from '../assets/assets.module';
 import { AccountsProvider } from '../account-stats/loaders/accounts.loader';
 import { AccountsRedisHandler } from '../account-stats/loaders/accounts.redis-handler';
@@ -8,12 +8,12 @@ import { CollectionAssetsProvider } from './loaders/collection-assets.loader';
 import { CollectionAssetsRedisHandler } from './loaders/collection-assets.redis-handler';
 import { CollectionAssetsCountProvider } from './loaders/collection-assets-count.loader';
 import { CollectionAssetsCountRedisHandler } from './loaders/collection-assets-count.redis-handler';
-import { CollectionAssetsResolver } from './collections-assets.resolver';
-import { CollectionsMutationsResolver } from './collections-mutations.resolver';
+import { CollectionAssetsResolver } from './collection-assets.resolver';
+import { CollectionsMutationsResolver } from './collection-mutations.resolver';
 import { CachingService } from 'src/common/services/caching/caching.service';
 import { LocalCacheService } from 'src/common/services/caching/local.cache.service';
-import { CollectionsNftsRedisHandler } from './collections-nfts.redis-handler';
-import { CollectionsNftsCountRedisHandler } from './collections-nfts-count.redis-handler';
+import { CollectionsNftsRedisHandler } from './collection-nfts.redis-handler';
+import { CollectionsNftsCountRedisHandler } from './collection-nfts-count.redis-handler';
 import { ElrondCommunicationModule } from 'src/common/services/elrond-communication/elrond-communication.module';
 import { AuctionsForCollectionProvider } from './loaders/collection-auctions.loader';
 import { AuctionsForCollectionRedisHandler } from './loaders/collection-auctions.redis-handler';
@@ -43,4 +43,4 @@ import { AuctionsForCollectionRedisHandler } from './loaders/collection-auctions
   ],
   exports: [CollectionsService, LocalCacheService],
 })
-export class CollectionsModuleGraph {}
+export class CollectionModuleGraph {}
